pyproject.toml
[tool.poetry]
name = "truetone-backend"
version = "0.1.0"
description = "TrueTone backend – API and ML worker services"
authors = ["TrueTone Team"]
readme = "README.md"
packages = [
    { include = "app", from = "api" },
    { include = "app", from = "worker" },
    { include = "shared" },
]

[tool.poetry.dependencies]
python = "^3.11"

# API
fastapi = "^0.111.0"
uvicorn = { version = "^0.29.0", extras = ["standard"] }
python-multipart = "^0.0.9"
pydantic = "^2.7.0"
pydantic-settings = "^2.2.1"

# AWS
boto3 = "^1.34.0"

# Worker / ML
mediapipe = "^0.10.14"
diffusers = "^0.27.2"
transformers = "^4.40.0"
accelerate = "^0.30.0"
torch = "^2.3.0"
Pillow = "^10.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.6"
httpx = "^0.27.0"
moto = { version = "^5.0.0", extras = ["s3", "dynamodb", "sqs"] }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
